requirements.txt
pandas==1.4.1
pymysql==1.0.2
sqlalchemy==1.4.32
python-dotenv==0.16.0
sqlalchemy-utils==0.38.3
kafka-python==2.0.2
requests==2.28.1
web3
aiohttp
yahoo-fin
hexbytes
azure-eventhub
azure-identity
azure-storage-blob
azure-data-tables
azure-keyvault-secrets